+++ test/mock/service.go
@@ -94,12 +94,12 @@ func (m *ServicePersistence) Count(metric string) (res int) {
 // ServiceHash
 type ServiceHash struct{}
 
-func (h *ServiceHash) Hash(a, b []byte) ([]byte, uint64) {
+func (h *ServiceHash) Hash(a, b []byte) ([]byte, uint16) {
 	sha := sha1.Sum(append(a, b...))
-	return sha[:], uint64(0)
+	return sha[:], uint16(0)
 }
-func (h *ServiceHash) GetPartition([]byte) uint64 {
-	return uint64(0)
+func (h *ServiceHash) GetPartition([]byte) uint16 {
+	return uint16(0)
 }
 
 // ServiceCoordination
